+++ CHANGELOG.md
@@ -19,6 +19,13 @@ The most recent changes are listed first.
 
 - Add the ability to not create and maintain a compact block cache.
 
+- Add support for transparent transactions.
+
+- Add `poolType` argument to `GetBlockRange` and `GetMempoolTx`. This
+  filtering allows the caller to request specific components (transparent,
+  shielded, or a combination) of blocks (`GetBlockRange`) and transactions
+  (`GetMempoolTx`).
+
 
 ### Changed